# backend/src/main/resources/static/api-docs.yaml
openapi: 3.0.0
info:
  title: Springyield API
  version: v0.1
servers:
  - url: http://springyield-springboot:8080
  - url: https://api.springyield.stefvisser.com
paths:
...
  /api/transactions/create:
    post:
      tags:
        - transaction-controller
      operationId: createTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/transactions/atm:
    post:
      tags:
        - transaction-controller
      operationId: processAtmTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/user/{userId}:
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/transactions/search:
    get:
      tags:
        - transaction-controller
      operationId: searchTransactions
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: amountFrom
          in: query
          required: false
          schema:
            type: number
        - name: amountTo
          in: query
          required: false
          schema:
            type: number
        - name: amountOperator
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/transactions/reference/{reference}:
    get:
      tags:
        - transaction-controller
      operationId: getTransactionsByReference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/transactions/id/{transactionID}:
    get:
      tags:
        - transaction-controller
      operationId: getTransactionByID
      parameters:
        - name: transactionID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/transactions/iban/{iban}:
    get:
      tags:
        - transaction-controller
      operationId: getTransactionsByIBAN
      parameters:
        - name: iban
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/account/iban/{iban}:
    get:
      tags:
        - account-controller
      operationId: getAccountByIban
      parameters:
        - name: iban
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    TransactionDTO:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
        fromAccount:
          type: string
        toAccount:
          type: string
        transferAmount:
          type: number
        description:
          type: string
        reference:
          type: string
        timestamp:
          type: string
          format: date-time
        transactionType:
          type: string
          enum:
            - TRANSFER
            - DEPOSIT
            - WITHDRAW
...